Removed: useless cast (it was causing a runtime check failure)
git-svn-id: http://dev.ryzom.com/svn/trunk@66 60ea55b9-4480-4529-b8ac-ed5291525b0f

+++ nel/src/misc/bitmap.cpp
@@ -1333,7 +1333,7 @@ bool CBitmap::decompressDXT5()
 			uint8 codeAlpha[16];
 			for(j=0; j<16; j++)
 			{
-				codeAlpha[j] = (uint8)bitsAlpha & 7;
+				codeAlpha[j] = (bitsAlpha & 7);
 				bitsAlpha>>=3;
 			}
